requirements.txt
python_telegram_bot==10.0.2
